Use int instead of double in Value

+++ src/main/java/fr/cnes/sonar/report/model/Value.java
@@ -28,14 +28,14 @@ public class Value {
     /**
      * Number of occurrences
      */
-    private double count;
+    private int count;
 
     /**
      * Complete constructor
      * @param pVal value to give to val
      * @param pCount value to give to count
      */
-    public Value(final String pVal, final double pCount) {
+    public Value(final String pVal, final int pCount) {
         this.val = pVal;
         this.count = pCount;
     }
@@ -60,7 +60,7 @@ public class Value {
      * Getter for count
      * @return count
      */
-    public double getCount() {
+    public int getCount() {
         return count;
     }
 
@@ -68,7 +68,7 @@ public class Value {
      * Setter to give to count
      * @param pCount count
      */
-    public void setCount(final double pCount) {
+    public void setCount(final int pCount) {
         this.count = pCount;
     }
 }
